split chat components into subcomponents
